Move PicklistPickController Test Store
