+++ tests/Feature/Http/Controllers/Api/Picklist/PicklistPickController/StoreTest.php
@@ -2,13 +2,32 @@
 
 namespace Tests\Feature\Http\Controllers\Api\Picklist\PicklistPickController;
 
+use App\Models\Order;
+use App\Models\OrderProduct;
+use App\User;
 use Tests\TestCase;
 
 class StoreTest extends TestCase
 {
     /** @test */
-    public function test_store_call_returns_ok()
+    public function store_returns_an_ok_response()
     {
-        $this->markTestSkipped();
+        $user = factory(User::class)->create();
+        $order = factory(Order::class)->create();
+        $orderProduct = factory(OrderProduct::class)->create(['order_id' => $order->id]);
+
+        $response = $this->actingAs($user, 'api')->postJson(route('picks.store'), [
+            'product_id'               => $orderProduct->product_id,
+            'quantity_picked'          => 1,
+            'quantity_skipped_picking' => 0,
+            'order_product_ids'        => [$orderProduct->id],
+        ]);
+
+        $response->assertOk();
+        $response->assertJsonStructure([
+            'data' => [
+                '*' => [],
+            ],
+        ]);
     }
 }
